backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
langchain==0.3.4
langchain-openai==0.2.2
openai==1.47.0
tiktoken==0.7.0
pdfminer.six==20231228
pytesseract==0.3.13
Pillow==10.4.0
requests==2.32.3
pytest==8.3.2
httpx==0.27.2
pgvector==0.2.5
email-validator==2.2.0
google-generativeai==0.7.2
